+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added `restrict_to_topics` argument to `record` method of `ROSCore`, allowing
   users to restrict bag recording to specific topics by supplying a regular
   expression in the same manner as the `rosbag` command line tool.
+* Improved documentation for ROSCore, its associated parts, and messages.
 
 
 # 1.3.0 (2020-06-02)